refactor: Move cmd/config/ -> pkg/config/
